requirements.txt
astroid==1.6.5
Django==2.0.7
django-ckeditor==5.5.0
django-extensions==2.0.7
django-js-asset==1.1.0
django-taggit==0.22.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
pylint==1.9.2
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
wrapt==1.10.11
